#146 task-photo で uploadImage でなく， addActivityをするように変更
