+++ src/main/java/org/magcruise/citywalk/model/input/DescriptionInput.java
@@ -2,21 +2,12 @@ package org.magcruise.citywalk.model.input;
 
 public class DescriptionInput extends Input {
 
-	private String value;
-
 	public DescriptionInput() {
+		super();
 	}
 
 	public DescriptionInput(String value) {
-		this.value = value;
-	}
-
-	public String getValue() {
-		return value;
-	}
-
-	public void setValue(String value) {
-		this.value = value;
+		super(value);
 	}
 
 }
