+++ src/main/java/org/magcruise/citywalk/model/input/Input.java
@@ -4,4 +4,21 @@ import org.nkjmlab.util.json.JsonObject;
 
 public class Input extends JsonObject<Input> {
 
+	private String value;
+
+	public Input() {
+	}
+
+	public Input(String value) {
+		this.value = value;
+	}
+
+	public String getValue() {
+		return value;
+	}
+
+	public void setValue(String value) {
+		this.value = value;
+	}
+
 }
