+++ src/main/java/org/magcruise/citywalk/model/input/PhotoInput.java
@@ -2,4 +2,12 @@ package org.magcruise.citywalk.model.input;
 
 public class PhotoInput extends Input {
 
+	public PhotoInput() {
+		super();
+	}
+
+	public PhotoInput(String value) {
+		super(value);
+	}
+
 }
