+++ src/main/java/org/magcruise/citywalk/model/input/PinInput.java
@@ -1,0 +1,13 @@
+package org.magcruise.citywalk.model.input;
+
+public class PinInput extends Input {
+
+	public PinInput() {
+		super();
+	}
+
+	public PinInput(String value) {
+		super(value);
+	}
+
+}
