+++ src/main/java/org/magcruise/citywalk/model/input/QrCodeInput.java
@@ -2,4 +2,12 @@ package org.magcruise.citywalk.model.input;
 
 public class QrCodeInput extends Input {
 
+	public QrCodeInput() {
+		super();
+	}
+
+	public QrCodeInput(String value) {
+		super(value);
+	}
+
 }
