+++ src/main/java/org/magcruise/citywalk/model/input/SelectionInput.java
@@ -2,21 +2,12 @@ package org.magcruise.citywalk.model.input;
 
 public class SelectionInput extends Input {
 
-	private String value;
-
 	public SelectionInput() {
+		super();
 	}
 
 	public SelectionInput(String value) {
-		this.value = value;
-	}
-
-	public String getValue() {
-		return value;
-	}
-
-	public void setValue(String value) {
-		this.value = value;
+		super(value);
 	}
 
 }
