+++ src/main/webapp/js/task-photo.js
@@ -9,16 +9,11 @@ $(function() {
 
   $("#loading").hide();
   $("#task-img").attr('src', task.imgSrc);
-  $("#btn-next").click(
-          function() {
-            $("#loading").fadeIn();
-            var imgData = $("#img-preview").attr('src');
-            new JsonRpcClient(new JsonRpcRequest(getBaseUrl(), "uploadImage",
-                    [getUserId(), imgData], function(data) {
-                      $("#loading").fadeOut();
-                      addActivity(task, "", true);
-                    })).rpc();
-          });
+  $("#btn-next").click(function() {
+    $("#loading").fadeIn();
+    var imgData = $("#img-preview").attr('src');
+    addActivity(task, imgData, true);
+  });
 });
 
 function handleFiles(files) {
